+++ src/elements/GeoTree/GeoTree.vue
@@ -30,11 +30,19 @@
           :expanded-categories="expandedCategories"
           :key-for-subcategory="keyForSubcategory"
           :description-icon="descriptionIcon"
+          :collapsed-icon="collapsedIcon"
+          :expanded-icon="expandedIcon"
           @check="handleCheckItem"
           @toggleExpand="handleToggleExpand"
         >
-          <template v-slot:trailingAccessoryAction>
-            <slot name="actionButton" />
+          <template
+            slot="trailingAccessoryAction"
+            slot-scope="{ item }"
+          >
+            <slot
+              name="actionButton"
+              :item="item"
+            />
           </template>
         </geo-tree-item>
       </ul>
@@ -149,6 +157,29 @@ export default {
      */
     dynamicExpandedCategories: {
       type: Object,
+      required: false
+    },
+    /**
+     * Optional Font Awesome 5 icon to use as collapsed icon
+     *
+     * See [vue-fontawesome](https://www.npmjs.com/package/@fortawesome/vue-fontawesome#explicit-prefix-note-the-vue-bind-shorthand-because-this-uses-an-array)
+     * for more info about this.
+     */
+    collapsedIcon: {
+      type: Array,
+      required: false,
+      default: function () {
+        return ['fal', 'chevron-right']
+      }
+    },
+    /**
+     * Optional Font Awesome 5 icon to use as expanded icon
+     *
+     * See [vue-fontawesome](https://www.npmjs.com/package/@fortawesome/vue-fontawesome#explicit-prefix-note-the-vue-bind-shorthand-because-this-uses-an-array)
+     * for more info about this.
+     */
+    expandedIcon: {
+      type: Array,
       required: false
     }
   },